Split off lib and bin, update deps
