+++ Cargo.toml
@@ -5,4 +5,4 @@ edition = "2021"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-winnow = "0.3.6"
+winnow = "0.4.1"
